refactor(sales) : 폴더명 수정
